--- src/main/java/InputView.java
@@ -1,8 +1,0 @@
-import java.util.Scanner;
-
-public class InputView {
-    public static String getInput() {
-        Scanner scanner = new Scanner(System.in);
-        return scanner.nextLine();
-    }
-}
