+++ src/main/java/ui/InputView.java
@@ -1,6 +1,6 @@
 package ui;
 
-import model.Lotto;
+import model.lotto.Lotto;
 
 import java.util.Scanner;
 
@@ -17,5 +17,10 @@ public class InputView {
         Scanner scanner = new Scanner(System.in);
         String input = scanner.nextLine();
         return new Lotto(input);
+    }
+
+    public static String getInput() {
+        Scanner scanner = new Scanner(System.in);
+        return scanner.nextLine();
     }
 }
