+++ src/test/java/ExpressionTest.java
@@ -1,3 +1,4 @@
+import model.calculate.Expression;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
