+++ src/test/java/OperatorTest.java
@@ -1,3 +1,4 @@
+import model.calculate.Operator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
