+++ src/test/java/model/LottoTest.java
@@ -1,5 +1,6 @@
 package model;
 
+import model.lotto.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
