+++ src/test/java/model/RankTest.java
@@ -1,5 +1,6 @@
 package model;
 
+import model.lotto.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
